secure delAllDb with pwd in conf

+++ app/controllers/dev/Util.scala
@@ -14,8 +14,8 @@ object Util extends Controller with Authentication {
 
   def testRes(service: String) = Action { implicit request =>
     Async {
-      Endpoints.getConfig(service).flatMap { config => 
-        Endpoints.genererUrl(service, Map.empty, None).map { url => 
+      Endpoints.getConfig(service).flatMap { config =>
+        Endpoints.genererUrl(service, Map.empty, None).map { url =>
           config.provider.fetch(url).get.map { response =>
             Ok(config.provider.resultAsJson(response))
           }
@@ -23,10 +23,11 @@ object Util extends Controller with Authentication {
       }.getOrElse(future(BadRequest("Service not found")))
     }
   }
-  
-  def deleteUser() = Authenticated { user => request => 
+
+  def deleteUser() = Authenticated { user =>
+    request =>
       Async {
-        UserDao.findOneById(user.accounts.head.id).map { user => 
+        UserDao.findOneById(user.accounts.head.id).map { user =>
           user.map { u =>
             UserDao.delete(u)
             Ok("user deleted")
@@ -34,35 +35,39 @@ object Util extends Controller with Authentication {
         }
       }
   }
-  
-  def deleteAllUsers() = Authenticated { user => request => 
-    UserDao.findAll.foreach { users => 
-      users foreach { user =>
-        UserDao.delete(user)
+
+  def deleteAllUsers(pwd: String) = Authenticated { user =>
+    request =>
+      import play.api.Play.current
+      if (pwd == current.configuration.getString("pwdDelAllDb").get) {
+        UserDao.findAll.foreach { users =>
+          users foreach { user =>
+            UserDao.delete(user)
+          }
+        }
       }
-    }
-    Ok("users deleted")
+      Ok("users deleted")
   }
-  
+
   def userDistantRes(providerName: String) = Action { implicit request =>
     Async {
       ProviderDispatcher(providerName).map { provider =>
-       provider.fetch(provider.getUserInfosUrl.get).get.map { response =>
-        Ok(response.json)
-       }
+        provider.fetch(provider.getUserInfosUrl.get).get.map { response =>
+          Ok(response.json)
+        }
       }.getOrElse(future(Ok("Service not found")))
     }
   }
-  
+
   def userRes(providerName: String) = Action { implicit request =>
     import scala.concurrent.ExecutionContext.Implicits.global
     Async {
-      ProviderDispatcher(providerName).map { provider => 
+      ProviderDispatcher(providerName).map { provider =>
         provider.getUser.map { response =>
           Ok(Json.toJson(response))
         }
       }.getOrElse(future(Ok("Provider not found")))
     }
   }
-  
+
 }
